telegram_bot/requirements.txt
aiogram==3.1.1
aiogram-dialog==2.1.0
alembic==1.12.0
annotated-types==0.5.0
APScheduler==3.10.4
asyncpg==0.28.0
environs==9.5.0
eyed3==0.9.7
Jinja2==3.1.2
loguru==0.7.2
Pillow==10.0.1
pytz==2023.3.post1
redis==5.0.1
selenium==4.14.0
SQLAlchemy==2.0.22
uvicorn==0.23.2
yt-dlp==2023.10.13